move test to integration test to avoid timeout
